soundfile_format reorg

+++ reinstall_linux.sh
@@ -10,11 +10,22 @@ fi
 
 # Build lib
 cd ./lib/build/linux
+if [ -d py$pyver ]; then
+    rm -r py$pyver
+fi
 ./build.sh $pyver
 cd ../../..
 
 installdir=$(python${pyver} -c "from distutils.sysconfig import get_python_lib; print(get_python_lib(prefix='/usr/local'))")
 
-sudo rm -r ${installdir}/medussa
-sudo rm ${installdir}/medussa-*.egg-info
+if [ -d ${installdir}/medussa ]; then
+    sudo rm -r ${installdir}/medussa
+fi
+if [ -e ${installdir}/medussa-*.egg-info ]; then
+    sudo rm ${installdir}/medussa-*.egg-info
+fi
+if [ -d build ]; then
+    sudo rm -r build
+fi
+python${pyver} setup.py build
 sudo python${pyver} setup.py install --prefix='/usr/local'
